requirements.txt
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
PyYAML==3.11
SQLAlchemy==1.0.4
alembic==0.7.6
boto==2.38.0
psycopg2==2.6
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
wsgiref==0.1.2
